Move readme up
I'm adding a capture template for new days which get appended to the
file. It would be nice to have the Readme heading at the bottom, but
this isn't trivial so I'm just doing this. It looks weird when you're
viewing it from GitHub but I don't care too much about that.