Modify day4 structure
I was close to giving up on literate programming because Org mode and
Clojure literate programming don't work perfectly together yet. The
previous solution worked if you did (ns aoc.4) in the REPL because the
ordering of functions wasn't as they would need to be. Doing that in
the REPL isn't feasible in the future, but reordering the functions
kind of limits the freedom literate programming gives.