Fix 2021.07
My old implementation had two bugs. First, it incorrectly did the
binary search jumping, jumping to (n cnt) instead of (n b). This
probably made the binary search do a lot more steps than necessary,
but wouldn't lead to an incorrect result.

The other bug could definitely result in incorrect results, but didn't
with this input. That bug was that we set b to (count vals), rather
than the maximum value. Imagine an input like this:

     99
     99
     99
     99

then b would be set to 4 (a to 0), and the binary search algorithm
would loop forever. Fortunately, the puzzle input was close enough to
0 on average and had enough elements that (count elements) was higher
than the goal.